refactor(shared): convert HTTP methods type to enum ml-61 (#124)
* feat: add http method enums ml-61

* chore: delete http method type ml-61

* refactor: change type using enums ml-61

* chore: add path alias for simplified imports ml-61

* fix: change to relative path ml-61
