+++ apps/backend/src/libs/constants/libs/types/white-route.type.ts
@@ -1,7 +1,8 @@
 import { type HTTPMethod } from "~/libs/modules/http/http.js";
+import { type ValueOf } from "~/libs/types/types.js";
 
 type WhiteRoute = {
-	method: HTTPMethod;
+	method: ValueOf<typeof HTTPMethod>;
 	path: string;
 };
