+++ apps/backend/src/libs/modules/controller/libs/types/controller-route-parameters.type.ts
@@ -1,13 +1,13 @@
 import { type preHandlerHookHandler } from "fastify";
 
 import { type HTTPMethod } from "~/libs/modules/http/http.js";
-import { type ValidationSchema } from "~/libs/types/types.js";
+import { type ValidationSchema, type ValueOf } from "~/libs/types/types.js";
 
 import { type APIHandler } from "./api-handler.type.js";
 
 type ControllerRouteParameters = {
 	handler: APIHandler;
-	method: HTTPMethod;
+	method: ValueOf<typeof HTTPMethod>;
 	path: string;
 	preHandlers?: preHandlerHookHandler | preHandlerHookHandler[] | undefined;
 	validation?: {
