+++ apps/backend/src/libs/modules/http/http.ts
@@ -1,3 +1,2 @@
-export { HTTPCode } from "./libs/enums/enums.js";
+export { HTTPCode, HTTPMethod } from "./libs/enums/enums.js";
 export { HTTPError } from "./libs/exceptions/exceptions.js";
-export { type HTTPMethod } from "./libs/types/types.js";
