+++ apps/backend/src/libs/modules/http/libs/enums/enums.ts
@@ -1,1 +1,1 @@
-export { HTTPCode } from "@meetlytic/shared";
+export { HTTPCode, HTTPMethod } from "@meetlytic/shared";
