--- apps/backend/src/libs/modules/http/libs/types/types.ts
@@ -1,1 +1,0 @@
-export { type HTTPMethod } from "@meetlytic/shared";
