+++ apps/backend/src/libs/modules/server-application/libs/types/server-application-route-parameters.type.ts
@@ -5,14 +5,14 @@ import {
 } from "fastify";
 
 import { type HTTPMethod } from "~/libs/modules/http/http.js";
-import { type ValidationSchema } from "~/libs/types/types.js";
+import { type ValidationSchema, type ValueOf } from "~/libs/types/types.js";
 
 type ServerApplicationRouteParameters = {
 	handler: (
 		request: FastifyRequest,
 		reply: FastifyReply,
 	) => Promise<void> | void;
-	method: HTTPMethod;
+	method: ValueOf<typeof HTTPMethod>;
 	path: string;
 	preHandlers?: preHandlerHookHandler | preHandlerHookHandler[] | undefined;
 	validation?: {
