+++ packages/shared/src/index.ts
@@ -15,10 +15,10 @@ export { configureString } from "./libs/helpers/helpers.js";
 export { type Config } from "./libs/modules/config/config.js";
 export {
 	type HTTP,
-	type HTTPMethod,
 	type HTTPOptions,
 	HTTPCode,
 	HTTPHeader,
+	HTTPMethod,
 } from "./libs/modules/http/http.js";
 export { type Storage } from "./libs/modules/storage/storage.js";
 export {
