+++ packages/shared/src/libs/modules/http/libs/enums/enums.ts
@@ -1,2 +1,3 @@
 export { HTTPCode } from "./http-code.enum.js";
 export { HTTPHeader } from "./http-header.enum.js";
+export { HTTPMethod } from "./http-method.enum.js";
