+++ packages/shared/src/libs/modules/http/libs/enums/http-method.enum.ts
@@ -1,0 +1,7 @@
+const HTTPMethod = {
+	DELETE: "DELETE",
+	GET: "GET",
+	POST: "POST",
+	PUT: "PATCH",
+};
+export { HTTPMethod };
