--- packages/shared/src/libs/modules/http/libs/types/http-method.type.ts
@@ -1,3 +1,0 @@
-type HTTPMethod = "DELETE" | "GET" | "PATCH" | "POST";
-
-export { type HTTPMethod };
