+++ packages/shared/src/libs/modules/http/libs/types/http-options.type.ts
@@ -1,8 +1,9 @@
-import { type HTTPMethod } from "./http-method.type.js";
+import { type ValueOf } from "../../../../types/types.js";
+import { type HTTPMethod } from "../enums/enums.js";
 
 type HTTPOptions = {
 	headers: Headers;
-	method: HTTPMethod;
+	method: ValueOf<typeof HTTPMethod>;
 	payload: BodyInit | null;
 };
