+++ packages/shared/src/libs/modules/http/libs/types/types.ts
@@ -1,3 +1,2 @@
-export { type HTTPMethod } from "./http-method.type.js";
 export { type HTTPOptions } from "./http-options.type.js";
 export { type HTTP } from "./http.type.js";
